requirements.txt
lxml==4.7.1
Pillow==9.0.1
python-docx==0.8.11
python-dotenv==0.19.2
qrcode==7.3.1
